+++ hahomematic/custom_platforms/switch.py
@@ -46,7 +46,7 @@ class CeSwitch(CustomEntity, OnTimeMixin):
             field_name=FIELD_CHANNEL_STATE, entity_type=HmBinarySensor
         )
 
-    @property
+    @value_property
     def channel_value(self) -> bool | None:
         """Return the current channel value of the switch."""
         return self._e_channel_state.value