+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 2023.2.8 (2023-02-09)
 
 - Add project setup script
+- Add entity_data event
+- Add payload mixin
 
 # Version 2023.2.7 (2023-02-07)